--- nodes/locales/en-US/ui-example.json
@@ -1,7 +1,0 @@
-{
-    "ui-example" : {
-        "label": {
-            "size": "Size"
-        }
-    }
-}
