+++ nodes/locales/en-US/ui-iframe.html
@@ -1,3 +1,3 @@
-<script type="text/html" data-help-name="ui-example">
+<script type="text/html" data-help-name="ui-iframe">
     <p>In-Editor help text will go here.</p>
 </script>
