+++ package.json
@@ -1,29 +1,23 @@
 {
-    "name": "@flowfuse/node-red-dashboard-2-ui-example",
+    "name": "@flowfuse/node-red-dashboard-2-ui-iframe",
     "version": "1.0.0",
-    "description": "A basic example of a third-party node that integrates with Node-RED Dashboard 2.0",
+    "description": "A Node-RED Dashboard 2.0 widget that allows you to embed an iframe into your dashboard",
     "keywords": [
         "node-red",
         "node-red-dashboard-2"
     ],
     "repository": {
         "type": "git",
-        "url": "https://github.com/flowforge/node-red-dashboard-3rd-party.git"
+        "url": "https://github.com/FlowFuse/node-red-dashboard-2-ui-iframe.git"
     },
     "license": "Apache-2.0",
     "author": {
         "name": "Joe Pavitt",
         "url": "https://github.com/joepavitt"
     },
-    "contributors": [
-        {
-            "name": "Pez Cuckow",
-            "url": "https://github.com/pezmc"
-        }
-    ],
     "exports": {
-        "import": "./resources/ui-example.esm.js",
-        "require": "./resources/ui-example.umd.js"
+        "import": "./resources/ui-iframe.esm.js",
+        "require": "./resources/ui-iframe.umd.js"
     },
     "files": [
         "dist/*",
@@ -64,15 +58,15 @@
     "node-red": {
         "version": ">=3.0.0",
         "nodes": {
-            "ui-example": "nodes/ui-example.js"
+            "ui-iframe": "nodes/ui-iframe.js"
         }
     },
     "node-red-dashboard-2": {
         "version": "1.0.0",
         "widgets": {
-            "ui-example": {
-                "output": "ui-example.umd.js",
-                "component": "UIExample"
+            "ui-iframe": {
+                "output": "ui-iframe.umd.js",
+                "component": "UIIframe"
             }
         }
     }
