+++ resources/ui-iframe.umd.js
@@ -1,0 +1,2 @@
+(function(){"use strict";try{if(typeof document<"u"){var e=document.createElement("style");e.appendChild(document.createTextNode(".ui-iframe[data-v-33c46a1b]{width:100%;height:100%}")),document.head.appendChild(e)}}catch(t){console.error("vite-plugin-css-injected-by-js",t)}})();
+(function(t,s){typeof exports=="object"&&typeof module<"u"?s(exports,require("vuex"),require("vue")):typeof define=="function"&&define.amd?define(["exports","vuex","vue"],s):(t=typeof globalThis<"u"?globalThis:t||self,s(t["ui-iframe"]={},t.vuex,t.Vue))})(this,function(t,s,d){"use strict";const c=(e,i)=>{const o=e.__vccOpts||e;for(const[n,r]of i)o[n]=r;return o},a={name:"UIIframe",inject:["$socket"],props:{id:{type:String,required:!0},props:{type:Object,default:()=>({})},state:{type:Object,default:()=>({enabled:!1,visible:!1})}},computed:{...s.mapState("data",["messages"]),src:function(){var e;return((e=this.messages[this.id])==null?void 0:e.src)||this.props.src}},mounted(){this.$socket.on("widget-load:"+this.id,e=>{this.$store.commit("data/bind",{widgetId:this.id,msg:e})}),this.$socket.on("msg-input:"+this.id,e=>{this.$store.commit("data/bind",{widgetId:this.id,msg:e})}),this.$socket.emit("widget-load",this.id)},unmounted(){var e,i;(e=this.$socket)==null||e.off("widget-load"+this.id),(i=this.$socket)==null||i.off("msg-input:"+this.id)}},u=["src"];function f(e,i,o,n,r,m){return d.openBlock(),d.createElementBlock("iframe",{src:m.src,class:"ui-iframe"},null,8,u)}const p=c(a,[["render",f],["__scopeId","data-v-33c46a1b"]]);t.UIIframe=p,Object.defineProperty(t,Symbol.toStringTag,{value:"Module"})});
