+++ ui/components/UIIframe.vue
@@ -1,0 +1,52 @@
+<template>
+    <iframe :src="src" class="ui-iframe" />
+</template>
+
+<script>
+import { mapState } from 'vuex'
+
+export default {
+    name: 'UIIframe',
+    inject: ['$socket'],
+    props: {
+        id: { type: String, required: true },
+        props: { type: Object, default: () => ({}) },
+        state: { type: Object, default: () => ({ enabled: false, visible: false }) }
+    },
+    computed: {
+        ...mapState('data', ['messages']),
+        src: function () {
+            return this.messages[this.id]?.src || this.props.src
+        }
+    },
+    mounted () {
+        this.$socket.on('widget-load:' + this.id, (msg) => {
+            // load the latest message from the Node-RED datastore when this widget is loaded
+            // storing it in our vuex store so that we have it saved as we navigate around
+            this.$store.commit('data/bind', {
+                widgetId: this.id,
+                msg
+            })
+        })
+        this.$socket.on('msg-input:' + this.id, (msg) => {
+            // store the latest message in our client-side vuex store when we receive a new message
+            this.$store.commit('data/bind', {
+                widgetId: this.id,
+                msg
+            })
+        })
+        // tell Node-RED that we're loading a new instance of this widget
+        this.$socket.emit('widget-load', this.id)
+    },
+    unmounted () {
+        /* Make sure, any events you subscribe to on SocketIO are unsubscribed to here */
+        this.$socket?.off('widget-load' + this.id)
+        this.$socket?.off('msg-input:' + this.id)
+    }
+}
+</script>
+
+<style scoped>
+/* CSS is auto scoped, but using named classes is still recommended */
+@import "../stylesheets/ui-iframe.css";
+</style>
