+++ ui/index.js
@@ -1,2 +1,2 @@
 /* Exports of all of the components in this node */
-export { default as UIExample } from './components/UIExample.vue'
+export { default as UIIframe } from './components/UIIframe.vue'
