+++ ui/stylesheets/ui-iframe.css
@@ -1,0 +1,4 @@
+.ui-iframe {
+    width: 100%;
+    height: 100%;
+}
